refactor: 移动 Panel 相关扩展方法
