+++ src/Extensions/Components/BootstrapBlazor.DockView/wwwroot/js/dockview-panel.js
@@ -37,4 +37,29 @@ const updateTitle = panel => {
     }
 }
 
-export { updateDockviewPanel };
+const getPanels = content => {
+    return getPanel(content[0])
+}
+
+const getPanel = (contentItem, parent = {}, panels = []) => {
+    if (contentItem.type === 'component') {
+        panels.push({
+            id: contentItem.id,
+            groupId: contentItem.groupId,
+            title: contentItem.title,
+            tabComponent: contentItem.componentName,
+            contentComponent: contentItem.componentName,
+            params: { ...contentItem, parentId: parent.id }
+        })
+    } else {
+        contentItem.content?.forEach(item => getPanel(item, contentItem, panels))
+    }
+    return panels
+}
+
+const findPanelFunc = v => p => findPanel(p, v);
+
+const findPanel = (p, v) => (p.params.key && p.params.key === v.params.key) || p.id === v.id || p.title === v.title;
+
+
+export { updateDockviewPanel, findPanelFunc, findPanel };
